backend/requirements/dev.txt
-r base.txt
django-debug-toolbar==4.3.0
ipython==8.25.0
